converstational_bot/requirements.txt
# Core dependencies for the Conversational Bot
pyaudio>=0.2.11
websocket-client>=1.6.0
requests>=2.28.0
murf>=1.0.0

# LLM Integration
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Web UI
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0

# Optional dependencies for better audio playback
pygame>=2.1.0

# Development and testing
pytest>=7.0.0
